requirements.txt
asgiref==3.9.1
asttokens==3.0.0
camelot-py==1.0.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
decorator==5.2.1
distro==1.9.0
Django==5.2.5
django-ckeditor==6.7.3
django-debug-toolbar==6.0.0
django-js-asset==3.1.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
ipython==8.37.0
jedi==0.19.2
lxml==6.0.0
matplotlib-inline==0.1.7
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tabula-py==2.10.0
tabulate==0.9.0
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
